Quality filtering
Committing code to perform quality filtering on the expression data
from Kee et al. In quality_filtering.m we read in the expression data,
and align it with the relevant metadata. Then we filter genes based on
expression levels and their differential expression with respect to the
biological conditions of interest. The filtered expression matrix and
the relevant metadata are exported, and can now be used for downstream
analysis and modeling.